pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "desafiolacrei"
version = "0.1.0"
description = ""
authors = [
    {name = "giocarvalho07",email = "giovannidecarvalhomendes@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
